VERSION 0.08 ; sowhat.pl
in sowh.pl - minimal changes
altered get_stats to provide delta and delta prime for sowhat test.
Does not calculate p value if reps=1
also changed it so that the number of reps is now correct (before was
not calcuating the first time, only doing 99)
p-value is printed from 2nd rep

Created sowhat.pl
This runs the sowh test with reps set to 1, extracts a distribution of
test statistics (delta) and null distribution statistics (delta primes)
and performs a KS test - returns a p-value as well as the variance of
the null distribution.
Also included an example command for this in the Examples.txt - TEST 7

Also added a bunch of data sets to the published data set folder