moved extraneous published data sets out of program, added analysis.output
analysis.output will include all of the command output from the
analyses used for the paper. This includes a README file which has all
of the commands used in the paper. This can be updated to include the
commands used to formulate the graphs, etc.
The published data sets now just includes the data sets used in the
paper (which are 3). The others are still on my computer, as well as in
past versions. I updated the README, though it still lacks information
on the Hox.