Version 0.10 - Removed KS test, adjusted output of report
Program now runs the sowhat option using a one tailed test with the
test statistic set to the mean of the distribution of test statistics
(delta). This better parallels the classic SOWH test, but does not take
advantage of all of the statistical tests available.
But, this version does print out both distributions in the
sowh.info.test file, allowing the user to take full advantage of the
information.
This version also prints out a small plot of the variance
(variance.png).