VERSION 0.11 - CHAINS, RECALCULATE
This version represents a change in the overall structure of the
program. The option sowhat has been replaced with two options -
recalculate, which has the test statistic recalculated each iteration,
and chains=some number, which runs multiple chains of the test.
Additionally, the names of the output files have been adjusted to match
these options.
The example command is the last command in Examples.txt